# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Project Apex API
  description: |
    Project Apex is a comprehensive AI agent orchestration platform for managing tasks, agents,
    DAGs (Directed Acyclic Graphs), and approval workflows.

    ## Overview

    The Apex API provides endpoints for:
    - **Tasks**: Create, manage, and monitor AI agent tasks
    - **Agents**: Register and manage AI agents in the system
    - **DAGs**: Define and execute complex multi-step workflows
    - **Approvals**: Human-in-the-loop approval workflows
    - **Metrics**: System and performance metrics
    - **Health**: Service health and readiness checks

    ## Authentication

    The API supports two authentication methods:
    - **Bearer Token**: Include `Authorization: Bearer <token>` header
    - **API Key**: Include `X-API-Key: <api-key>` header

    ## Rate Limiting

    Rate limits are enforced based on your subscription tier:
    - **Standard**: 100 requests/minute
    - **Pro**: 1,000 requests/minute
    - **Enterprise**: Unlimited

    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets

    ## WebSocket

    Real-time updates are available via WebSocket at `wss://api.apex.io/ws`.
    See the WebSocket section for details on subscribing to events.

  version: 1.0.0
  contact:
    name: Apex API Support
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: List tasks
      description: Retrieve a paginated list of tasks with optional filtering.
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: status
          in: query
          description: Filter by task status
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: agentId
          in: query
          description: Filter by assigned agent ID
          schema:
            type: string
            format: uuid
        - name: dagId
          in: query
          description: Filter by parent DAG ID
          schema:
            type: string
            format: uuid
        - name: priority
          in: query
          description: Filter by task priority
          schema:
            $ref: '#/components/schemas/TaskPriority'
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
          example: "research,ai"
        - name: createdAfter
          in: query
          description: Filter tasks created after this timestamp
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          description: Filter tasks created before this timestamp
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
              example:
                items:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Research AI Trends"
                    description: "Research the latest trends in AI agents"
                    status: "running"
                    priority: "normal"
                    agentId: "agent-001"
                    retries: 3
                    retryCount: 0
                    tags: ["research", "ai"]
                    metadata: {}
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-15T10:30:05Z"
                    startedAt: "2024-01-15T10:30:05Z"
                total: 150
                page: 1
                perPage: 20
                totalPages: 8
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Tasks
      summary: Create a task
      description: Create and queue a new task for execution.
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            example:
              name: "Research AI Trends"
              description: "Research the latest trends in AI agents and summarize key findings"
              priority: "normal"
              agentId: "agent-001"
              input:
                data:
                  domain: "technology"
                  depth: "comprehensive"
                parameters:
                  output_format: "markdown"
                  max_sources: 10
              timeoutSeconds: 300
              retries: 3
              tags: ["research", "ai"]
              metadata:
                requestedBy: "user@example.com"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                name: "Research AI Trends"
                description: "Research the latest trends in AI agents and summarize key findings"
                status: "pending"
                priority: "normal"
                agentId: "agent-001"
                input:
                  data:
                    domain: "technology"
                    depth: "comprehensive"
                  parameters:
                    output_format: "markdown"
                    max_sources: 10
                timeoutSeconds: 300
                retries: 3
                retryCount: 0
                tags: ["research", "ai"]
                metadata:
                  requestedBy: "user@example.com"
                createdAt: "2024-01-15T10:30:00Z"
                updatedAt: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get a task
      description: Retrieve detailed information about a specific task.
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                name: "Research AI Trends"
                description: "Research the latest trends in AI agents"
                status: "completed"
                priority: "normal"
                agentId: "agent-001"
                input:
                  data:
                    domain: "technology"
                output:
                  result: "## AI Agent Trends 2024\n\n1. Multi-agent systems..."
                  artifacts: ["report.md", "sources.json"]
                  metrics:
                    tokens_used: 8542
                    sources_analyzed: 12
                timeoutSeconds: 300
                retries: 3
                retryCount: 0
                tags: ["research", "ai"]
                metadata: {}
                createdAt: "2024-01-15T10:30:00Z"
                updatedAt: "2024-01-15T10:32:15Z"
                startedAt: "2024-01-15T10:30:05Z"
                completedAt: "2024-01-15T10:32:15Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Tasks
      summary: Update a task
      description: Update an existing task. Only pending tasks can be fully updated.
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            example:
              name: "Updated Research Task"
              priority: "high"
              tags: ["research", "ai", "urgent"]
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: Delete a task. Only pending or completed tasks can be deleted.
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/{taskId}/cancel:
    post:
      tags:
        - Tasks
      summary: Cancel a task
      description: Cancel a pending or running task.
      operationId: cancelTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for cancellation
                  maxLength: 500
              example:
                reason: "No longer needed"
      responses:
        '200':
          description: Task cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                name: "Research AI Trends"
                status: "cancelled"
                priority: "normal"
                createdAt: "2024-01-15T10:30:00Z"
                updatedAt: "2024-01-15T10:31:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Task cannot be cancelled (already completed or cancelled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /agents:
    get:
      tags:
        - Agents
      summary: List agents
      description: Retrieve a paginated list of registered agents.
      operationId: listAgents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: status
          in: query
          description: Filter by agent status
          schema:
            $ref: '#/components/schemas/AgentStatus'
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
        - name: capability
          in: query
          description: Filter by capability name
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentList'
              example:
                items:
                  - id: "agent-001"
                    name: "Research Agent"
                    description: "Specialized in research tasks"
                    status: "busy"
                    capabilities:
                      - name: "web_search"
                        version: "1.0"
                      - name: "text_analysis"
                        version: "2.0"
                    maxConcurrentTasks: 5
                    currentTasks: 3
                    totalTasksCompleted: 150
                    tags: ["research", "general"]
                    metadata: {}
                    lastHeartbeat: "2024-01-15T10:30:00Z"
                    createdAt: "2024-01-01T00:00:00Z"
                    updatedAt: "2024-01-15T10:30:00Z"
                total: 25
                page: 1
                perPage: 20
                totalPages: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Agents
      summary: Register an agent
      description: Register a new agent in the system.
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
            example:
              name: "Research Agent"
              description: "Specialized agent for research and analysis tasks"
              capabilities:
                - name: "web_search"
                  version: "1.0"
                  parameters:
                    max_results: 100
                - name: "text_analysis"
                  version: "2.0"
              maxConcurrentTasks: 5
              tags: ["research", "analysis"]
              metadata:
                model: "gpt-4o"
                region: "us-west-2"
      responses:
        '201':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Get an agent
      description: Retrieve detailed information about a specific agent.
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
              example:
                id: "agent-001"
                name: "Research Agent"
                description: "Specialized in research tasks"
                status: "busy"
                capabilities:
                  - name: "web_search"
                    version: "1.0"
                  - name: "text_analysis"
                    version: "2.0"
                maxConcurrentTasks: 5
                currentTasks: 3
                totalTasksCompleted: 150
                tags: ["research", "general"]
                metadata:
                  model: "gpt-4o"
                lastHeartbeat: "2024-01-15T10:30:00Z"
                createdAt: "2024-01-01T00:00:00Z"
                updatedAt: "2024-01-15T10:30:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Agents
      summary: Update an agent
      description: Update an existing agent's configuration.
      operationId: updateAgent
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
            example:
              name: "Updated Research Agent"
              maxConcurrentTasks: 10
              tags: ["research", "priority"]
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Agents
      summary: Delete an agent
      description: Remove an agent from the system. Agent must be idle with no assigned tasks.
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '204':
          description: Agent deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Agent cannot be deleted (has active tasks)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.

        Include the token in the Authorization header:
        ```
        Authorization: Bearer <your-token>
        ```

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key authentication.

        Include your API key in the X-API-Key header:
        ```
        X-API-Key: <your-api-key>
        ```

  # --------------------------------------------------------------------------
  # Parameters
  # --------------------------------------------------------------------------
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: perPage
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        enum: [createdAt, updatedAt, name, status, priority]
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: asc

  # --------------------------------------------------------------------------
  # Schemas
  # --------------------------------------------------------------------------
  schemas:
...
    TaskStatus:
      type: string
      enum: [pending, queued, running, paused, completed, failed, cancelled]
      description: Task execution status
    TaskPriority:
      type: string
      enum: [low, normal, high, critical]
      description: Task priority level
    AgentStatus:
      type: string
      enum: [idle, busy, offline, error]
      description: Agent status
...
    TaskList:
      type: object
      required:
        - items
        - total
        - page
        - perPage
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total:
          type: integer
          description: Total number of items
        page:
          type: integer
          description: Current page number
        perPage:
          type: integer
          description: Items per page
        totalPages:
          type: integer
          description: Total number of pages

    # Agent Schemas
...
  responses:
    BadRequest:
      description: Bad request - Invalid parameters or request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid request parameters"
              details:
                field: "name"
                reason: "must not be empty"
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or missing authentication credentials"
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when the rate limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Too many requests. Please retry after 60 seconds."
              details:
                limit: 100
                remaining: 0
                resetAt: "2024-01-15T10:31:00Z"
...