# src/backend/core/src/api/v2/mod.rs
//! V2 API module for Apex Core.
//!
//! This module contains the V2 API endpoints which are currently in preview.
//! V2 introduces the following changes from V1:
//!
//! - Enhanced error responses with more detailed context
//! - Streaming support for task execution
//! - Batch operations for tasks and agents
//! - GraphQL endpoint support (planned)
//! - Enhanced pagination with cursor-based navigation
//! - Improved filtering and sorting options
//!
//! **WARNING**: V2 is in preview and may change without notice.
//! Do not use in production until it becomes stable.

use axum::{
    extract::{Query, State},
    response::IntoResponse,
    routing::{delete, get, post},
    Json, Router,
};
use serde::{Deserialize, Serialize};
use uuid::Uuid;

use crate::api::{handlers, AppState};
use crate::dag::{Task, TaskInput, TaskStatus, TaskId};

/// V2 API prefix.
pub const V2_PREFIX: &str = "/api/v2";

/// Build the V2 API router.
///
/// V2 includes all V1 endpoints plus new features.
/// Some endpoints have enhanced functionality.
pub fn v2_router() -> Router<AppState> {
    Router::new()
        // Task endpoints (same as V1 for now)
        .route("/tasks", post(handlers::create_task))
        .route("/tasks", get(list_tasks_v2))
        .route("/tasks/:id", get(handlers::get_task))
        .route("/tasks/:id/status", get(handlers::get_task_status))
        .route("/tasks/:id/cancel", post(handlers::cancel_task))
        // Batch operations (V2 only)
        .route("/tasks/batch", post(batch_create_tasks))
        .route("/tasks/batch/cancel", post(batch_cancel_tasks))
        // DAG endpoints (same as V1 for now)
        .route("/dags", post(handlers::create_dag))
        .route("/dags/:id", get(handlers::get_dag))
        .route("/dags/:id/execute", post(handlers::execute_dag))
        .route("/dags/:id/status", get(handlers::get_dag_status))
        // Agent endpoints (same as V1 for now)
        .route("/agents", get(handlers::list_agents))
        .route("/agents", post(handlers::register_agent))
        .route("/agents/:id", get(handlers::get_agent))
        .route("/agents/:id", delete(handlers::remove_agent))
        .route("/agents/:id/stats", get(handlers::get_agent_stats))
        // Contract endpoints (same as V1 for now)
        .route("/contracts", get(handlers::list_contracts))
        .route("/contracts/:id", get(handlers::get_contract))
        // Stats
        .route("/stats", get(handlers::get_system_stats))
        // V2 specific: Version info
        .route("/version", get(version_info))
}

// ═══════════════════════════════════════════════════════════════════════════════
// V2 Specific Types
// ═══════════════════════════════════════════════════════════════════════════════

/// Pagination parameters for V2 endpoints.
#[derive(Debug, Deserialize)]
pub struct PaginationParams {
    /// Cursor for pagination (base64 encoded).
    pub cursor: Option<String>,
    /// Number of items per page (default: 20, max: 100).
    #[serde(default = "default_limit")]
    pub limit: u32,
    /// Sort field.
    pub sort_by: Option<String>,
    /// Sort direction.
    #[serde(default)]
    pub sort_order: SortOrder,
}

fn default_limit() -> u32 {
    20
}

#[derive(Debug, Default, Deserialize)]
#[serde(rename_all = "lowercase")]
pub enum SortOrder {
    #[default]
    Asc,
    Desc,
}

/// Paginated response wrapper for V2.
#[derive(Debug, Serialize)]
pub struct PaginatedResponse<T> {
    pub success: bool,
    pub data: Vec<T>,
    pub pagination: PaginationInfo,
}

#[derive(Debug, Serialize)]
pub struct PaginationInfo {
    pub total: u64,
    pub limit: u32,
    pub has_more: bool,
    pub next_cursor: Option<String>,
    pub prev_cursor: Option<String>,
}

/// Batch operation request.
#[derive(Debug, Deserialize)]
pub struct BatchRequest<T> {
    pub items: Vec<T>,
}

/// Batch operation response.
#[derive(Debug, Serialize)]
pub struct BatchResponse<T> {
    pub success: bool,
    pub results: Vec<BatchResult<T>>,
    pub summary: BatchSummary,
}

#[derive(Debug, Serialize)]
pub struct BatchResult<T> {
    pub index: usize,
    pub success: bool,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub data: Option<T>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub error: Option<String>,
}

#[derive(Debug, Serialize)]
pub struct BatchSummary {
    pub total: usize,
    pub succeeded: usize,
    pub failed: usize,
}

// ═══════════════════════════════════════════════════════════════════════════════
// V2 Handlers
// ═══════════════════════════════════════════════════════════════════════════════

/// List tasks with V2 cursor-based pagination.
pub async fn list_tasks_v2(
    State(state): State<AppState>,
    Query(params): Query<PaginationParams>,
) -> impl IntoResponse {
    let limit = params.limit.min(100) as i64;
    let offset = if let Some(ref cursor_str) = params.cursor {
        // Decode cursor to get the offset
        match base64_decode_offset(cursor_str) {
            Ok(off) => off,
            Err(_) => 0i64,
        }
    } else {
        0i64
    };

    let total = match state.db.get_task_count().await {
        Ok(count) => count as u64,
        Err(_) => return Json(PaginatedResponse::<serde_json::Value> {
            success: false,
            data: vec![],
            pagination: PaginationInfo {
                total: 0,
                limit: params.limit,
                has_more: false,
                next_cursor: None,
                prev_cursor: None,
            },
        }),
    };

    match state.db.get_tasks_paginated(limit + 1, offset).await {
        Ok(tasks) => {
            let has_more = tasks.len() as i64 > limit;
            let tasks: Vec<serde_json::Value> = tasks.iter().take(limit as usize).map(|t| {
                serde_json::json!({
                    "id": t.id,
                    "dag_id": t.dag_id,
                    "name": t.name,
                    "status": t.status,
                    "tokens_used": t.tokens_used,
                    "cost_dollars": t.cost_dollars,
                    "created_at": t.created_at.to_rfc3339(),
                    "started_at": t.started_at.map(|ts| ts.to_rfc3339()),
                    "completed_at": t.completed_at.map(|ts| ts.to_rfc3339()),
                })
            }).collect();

            let next_cursor = if has_more {
                Some(base64_encode_offset(offset + limit))
            } else {
                None
            };
            let prev_cursor = if offset > 0 {
                Some(base64_encode_offset((offset - limit).max(0)))
            } else {
                None
            };

            Json(PaginatedResponse::<serde_json::Value> {
                success: true,
                data: tasks,
                pagination: PaginationInfo {
                    total,
                    limit: params.limit,
                    has_more,
                    next_cursor,
                    prev_cursor,
                },
            })
        }
        Err(_) => Json(PaginatedResponse::<serde_json::Value> {
            success: false,
            data: vec![],
            pagination: PaginationInfo {
                total: 0,
                limit: params.limit,
                has_more: false,
                next_cursor: None,
                prev_cursor: None,
            },
        }),
    }
}

/// Encode an offset into a base64 cursor string.
fn base64_encode_offset(offset: i64) -> String {
    use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine};
    URL_SAFE_NO_PAD.encode(format!("offset:{}", offset).as_bytes())
}

/// Decode a base64 cursor string into an offset.
fn base64_decode_offset(cursor: &str) -> Result<i64, ()> {
    use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine};
    let bytes = URL_SAFE_NO_PAD.decode(cursor).map_err(|_| ())?;
    let s = String::from_utf8(bytes).map_err(|_| ())?;
    let offset_str = s.strip_prefix("offset:").ok_or(())?;
    offset_str.parse::<i64>().map_err(|_| ())
}

/// Batch create tasks.
pub async fn batch_create_tasks(
    State(_state): State<AppState>,
    Json(req): Json<BatchRequest<handlers::CreateTaskRequest>>,
) -> impl IntoResponse {
    let mut results: Vec<BatchResult<serde_json::Value>> = Vec::with_capacity(req.items.len());
    let mut succeeded = 0usize;
    let failed = 0usize;

    for (i, task_req) in req.items.iter().enumerate() {
        let input = TaskInput {
            instruction: task_req.instruction.clone(),
            context: task_req.context.clone().unwrap_or(serde_json::Value::Null),
            parameters: serde_json::Value::Null,
            artifacts: vec![],
        };

        let mut task = Task::new(&task_req.name, input);
        if let Some(priority) = task_req.priority {
            task.priority = priority;
        }

        let response = serde_json::json!({
            "id": task.id.0,
            "name": task.name,
            "status": task.status.as_str(),
            "tokens_used": task.tokens_used,
            "cost_dollars": task.cost_dollars,
            "created_at": task.created_at.to_rfc3339(),
        });

        results.push(BatchResult {
            index: i,
            success: true,
            data: Some(response),
            error: None,
        });
        succeeded += 1;
    }

    Json(BatchResponse {
        success: failed == 0,
        results,
        summary: BatchSummary {
            total: req.items.len(),
            succeeded,
            failed,
        },
    })
}

/// Batch cancel tasks.
pub async fn batch_cancel_tasks(
    State(state): State<AppState>,
    Json(req): Json<BatchRequest<Uuid>>,
) -> impl IntoResponse {
    let mut results: Vec<BatchResult<serde_json::Value>> = Vec::with_capacity(req.items.len());
    let mut succeeded = 0usize;
    let mut failed = 0usize;

    for (i, task_id) in req.items.iter().enumerate() {
        match state.db.update_task_status(TaskId(*task_id), TaskStatus::Cancelled).await {
            Ok(_) => {
                results.push(BatchResult {
                    index: i,
                    success: true,
                    data: Some(serde_json::json!({
                        "id": task_id,
                        "status": "cancelled",
                    })),
                    error: None,
                });
                succeeded += 1;
            }
            Err(e) => {
                results.push(BatchResult {
                    index: i,
                    success: false,
                    data: None,
                    error: Some(e.user_message().to_string()),
                });
                failed += 1;
            }
        }
    }

    Json(BatchResponse {
        success: failed == 0,
        results,
        summary: BatchSummary {
            total: req.items.len(),
            succeeded,
            failed,
        },
    })
}

/// Get V2 version information.
pub async fn version_info() -> impl IntoResponse {
    Json(serde_json::json!({
        "success": true,
        "data": {
            "api_version": "2.0.0",
            "status": "preview",
            "features": [
                "batch_operations",
                "cursor_pagination",
                "enhanced_errors"
            ],
            "deprecation_notice": "V2 is in preview. Do not use in production.",
            "stable_version": "1.0.0"
        }
    }))
}

/// V2 API route constants for use in clients and documentation.
pub mod routes {
    // Task routes
    pub const TASKS: &str = "/api/v2/tasks";
    pub const TASK: &str = "/api/v2/tasks/:id";
    pub const TASK_STATUS: &str = "/api/v2/tasks/:id/status";
    pub const TASK_CANCEL: &str = "/api/v2/tasks/:id/cancel";
    pub const TASKS_BATCH: &str = "/api/v2/tasks/batch";
    pub const TASKS_BATCH_CANCEL: &str = "/api/v2/tasks/batch/cancel";

    // DAG routes
    pub const DAGS: &str = "/api/v2/dags";
    pub const DAG: &str = "/api/v2/dags/:id";
    pub const DAG_EXECUTE: &str = "/api/v2/dags/:id/execute";
    pub const DAG_STATUS: &str = "/api/v2/dags/:id/status";

    // Agent routes
    pub const AGENTS: &str = "/api/v2/agents";
    pub const AGENT: &str = "/api/v2/agents/:id";
    pub const AGENT_STATS: &str = "/api/v2/agents/:id/stats";

    // Contract routes
    pub const CONTRACTS: &str = "/api/v2/contracts";
    pub const CONTRACT: &str = "/api/v2/contracts/:id";

    // System routes
    pub const STATS: &str = "/api/v2/stats";
    pub const VERSION: &str = "/api/v2/version";
}

#[cfg(test)]
mod tests {
    use super::*;

    #[test]
    fn test_route_constants() {
        assert!(routes::TASKS.starts_with("/api/v2"));
        assert!(routes::TASKS_BATCH.contains("batch"));
    }

    #[test]
    fn test_pagination_defaults() {
        let params: PaginationParams = serde_json::from_str("{}").unwrap();
        assert_eq!(params.limit, 20);
    }
}